--- a/src/daft-shuffles/src/lib.rs
+++ b/src/daft-shuffles/src/lib.rs
@@ -1,4 +1,4 @@
-pub mod shuffle_cache;
-
 #[cfg(feature = "python")]
 pub mod python;
+pub mod server;
+pub mod shuffle_cache;