--- a/src/daft-shuffles/src/server/mod.rs
+++ b/src/daft-shuffles/src/server/mod.rs
@@ -1,0 +1,2 @@
+pub mod server;
+mod stream;