--- a/src/daft-shuffles/src/server/mod.rs
+++ b/src/daft-shuffles/src/server/mod.rs
@@ -1,2 +1,2 @@
-pub mod server;
+pub mod flight_server;
 mod stream;