--- a/src/generated/spark-connect/Cargo.toml
+++ b/src/generated/spark-connect/Cargo.toml
@@ -1,7 +1,7 @@
 [dependencies]
 prost = "0.13.3"
 prost-types = "0.13.3"
-tonic = "0.12.3"
+tonic = {workspace = true}
 
 [lints]
 workspace = true