## Summary

Moves the flight server for the flight shuffle from python to rust.

Previously we used PyArrow's flight implementation, but for better control / performance we can implement our own in rust.

One notable optimization we can do here is instead of reading shuffle files into arrow arrays before sending them over rpc, we can read and send the raw bytes from the files because they are already serialized into arrow ipc stream format.

### Results

Question | Before (s) | After (s) | Difference (s) | % Change
-- | -- | -- | -- | --
1 | 31.58 | 30.30 | -1.28 | -4.05%
2 | 43.39 | 38.85 | -4.54 | -10.46%
3 | 35.78 | 32.75 | -3.03 | -8.47%
4 | 23.63 | 22.45 | -1.18 | -4.99%
5 | 76.53 | 76.17 | -0.36 | -0.47%
6 | 10.86 | 10.47 | -0.39 | -3.59%
7 | 43.65 | 42.78 | -0.87 | -1.99%
8 | 80.40 | 78.61 | -1.79 | -2.23%
9 | 115.04 | 111.32 | -3.72 | -3.23%
10 | 40.85 | 43.92 | 3.07 | 7.52%
11 | 27.88 | 22.91 | -4.97 | -17.83%
12 | 24.63 | 25.28 | 0.65 | 2.64%
13 | 33.76 | 31.59 | -2.17 | -6.43%
14 | 17.18 | 17.70 | 0.52 | 3.03%
15 | 38.31 | 39.52 | 1.21 | 3.16%
16 | 15.38 | 17.39 | 2.01 | 13.07%
17 | 113.43 | 112.06 | -1.37 | -1.21%
18 | 62.07 | 61.14 | -0.93 | -1.50%
19 | 22.01 | 21.88 | -0.13 | -0.59%
20 | 51.93 | 51.27 | -0.66 | -1.27%
21 | 134.14 | 135.87 | 1.73 | 1.29%
22 | 12.98 | 12.37 | -0.61 | -4.70%
Total | 1055.41 | 1036.60 | -18.81 | -1.78%

No noticeable changes across the board. The performance gains will likely come from implementing in-memory cache and optimizing the reduce side.

## Related Issues

None

## Changes Made

Implement a rust based flight server

## Checklist

- [x] All tests have passed
- [ ] Documented in API Docs
- [ ] Documented in User Guide
- [ ] If adding a new documentation page, doc is added to `docs/mkdocs.yml` navigation
- [ ] Documentation builds and is formatted properly (tag [@ccmao1130](https://github.com/ccmao1130) for docs review)